Commit draft captions automatically
I always forget to do it. I want to keep the draft in the git history.
It is tedious to rewind my text editor or regenerate the drafts after.
I am unlikely to interfere with a draft commit in progress, but if I
do I can undo that more quickly than redoing draft captions.