Change how backup works to account for the SSD
It is almost impossible to edit 4k footage accessed from a spinning
disk without tedious pauses, so I copy it to an SSD for speed.

This changes the backup to take this into account, backing up the SSD
to the spinning disk, that to another disk attached to my backup server
(raspberry pi), and that backup to S3.

Once the video is finalised I can remove the footage from the SSD
and relink the timeline in Resolve and still be able to use it.